implement: tailwind css from components
