+++ .prettierrc.json
@@ -9,5 +9,8 @@
   "trailingComma": "es5",
   "bracketSpacing": true,
   "arrowParens": "always",
-  "endOfLine": "lf"
+  "endOfLine": "lf",
+  "plugins": ["prettier-plugin-tailwindcss"],
+  "tailwindAttributes": ["twClassName"],
+  "tailwindFunctions": ["clsx"]
 }
