+++ package.json
@@ -32,6 +32,8 @@
     "eslint": "^8",
     "eslint-plugin-react": "^7.34.1",
     "postcss": "^8.4.38",
+    "prettier": "^3.2.5",
+    "prettier-plugin-tailwindcss": "^0.5.13",
     "tailwindcss": "^3.4.3",
     "typescript": "^5"
   }
