+++ src/app/[locale]/globals.css
@@ -8,6 +8,7 @@
   --background-start-rgb: 214, 219, 220;
   --background-end-rgb: 255, 255, 255;
 
+  --header-static-height: 5rem;
   --primary-color: #80bad6;
   --text-color: #a9a9a9;
   --text-light-color: rgb(210, 210, 210);
