+++ src/app/[locale]/page.module.css
@@ -1,11 +1,8 @@
 .main {
-  height: 100vh;
-  display: grid;
-  place-items: center;
-  overflow: hidden;
+  margin-top: var(--header-static-height);
 }
 
-.section {
+/* .section {
   position: relative;
   width: 100%;
   height: 100%;
@@ -18,7 +15,9 @@
   top: 50%;
   left: 3rem;
   transform: translateY(-50%);
-  font: 400 0.85rem helvetica, sans-serif;
+  font:
+    400 0.85rem helvetica,
+    sans-serif;
   color: white;
   text-shadow: 0 3px 8px rgba(0, 0, 0, 0.5);
   opacity: 0;
@@ -41,7 +40,9 @@
   padding: 0.75rem !important;
   cursor: pointer;
 
-  transition: border-color 0.3s ease, color 0.3s ease !important;
+  transition:
+    border-color 0.3s ease,
+    color 0.3s ease !important;
 }
 
 .read_button:hover {
@@ -71,10 +72,10 @@
 }
 .nav button:hover {
   background-color: rgba(255, 255, 255, 0.3);
-}
+} */
 
 /* Mobile */
-@media (max-width: 700px) {
+/* @media (max-width: 700px) {
   .title {
     font-size: 0.9rem;
   }
@@ -84,10 +85,10 @@
   .content button {
     font-size: 0.7rem;
   }
-}
+} */
 
 /* Tablet and Smaller Desktop */
-@media (min-width: 701px) and (max-width: 1120px) {
+/* @media (min-width: 701px) and (max-width: 1120px) {
   .title {
     font-size: 1rem;
   }
@@ -117,7 +118,7 @@
     left: calc(50% * 510px);
     opacity: 0;
   }
-}
+} */
 
 /* Dark Theme */
 /* @media (prefers-color-scheme: dark) {
